Change directory structure and edit setup.py accordingly
