+++ MANIFEST.in
@@ -1,0 +1,3 @@
+include *.fits
+include *.csv
+include *.md
