+++ slrealizer/utils/__init__.py
@@ -1,11 +1,2 @@
-#from SLRealizer import *
-#from distance import *
 from constants import *
-#from null_deblend import *
-#from catalog import *
-#from test import *
-#from extract_corner import *
-#from dropbox_manage import *
-#from format_sdss_to_om10 import *
-#from plotting import *
 from utils import *
